fix: move  check-block to services
