+++ wallet/src/nightfall-browser/services/transfer.js
@@ -11,7 +11,7 @@ It is agnostic to whether we are dealing with an ERC20 or ERC721 (or ERC1155).
 import gen from 'general-number';
 import { initialize } from 'zokrates-js';
 
-import confirmBlock from '../../utils/lib/check-block';
+import confirmBlock from './check-block';
 import computeCircuitInputs from '../utils/compute-witness';
 import getCommitmentInfo from '../utils/getCommitmentInfo';
 import { getContractInstance } from '../../common-files/utils/contract';
